initial pass at SPOpt class
