+++ mpisppy/spbase.py
@@ -17,7 +17,7 @@ logger = logging.getLogger("SPBase")
 logger.setLevel(logging.WARN)
 
 
-class SPBase(object):
+class SPBase:
     """ Defines an interface to all strata (hubs and spokes)
 
         Args:
@@ -465,8 +465,6 @@ class SPBase(object):
                                            " which are not 1")
                     checked_nodes.append(ndn)
 
-    # TODO: There is an issue (#60) to put these things on a block, but really we should
-    # have two blocks: one for Pyomo objects and the other for lists and caches.
     def _look_before_leap(self, scen, addlist):
         """ utility to check before attaching something to the user's model
         """
@@ -512,67 +510,3 @@ class SPBase(object):
             self._spcomm = weakref.ref(value)
         else:
             raise RuntimeError("SPBase.spcomm should only be set once")
-
-    def gather_var_values_to_rank0(self, get_zero_prob_values=False):
-        """ Gather the values of the nonanticipative variables to the root of
-        the `mpicomm` for the cylinder
-
-        Returns:
-            dict or None:
-                On the root (rank0), returns a dictionary mapping
-                (scenario_name, variable_name) pairs to their values. On other
-                ranks, returns None.
-        """
-        var_values = dict()
-        for (sname, model) in self.local_scenarios.items():
-            for node in model._mpisppy_node_list:
-                for var in node.nonant_vardata_list:
-                    if (self.is_zero_prob(model, var)) and (not get_zero_prob_values):
-                        var_values[sname, var.name] = None
-                    else:
-                        var_values[sname, var.name] = pyo.value(var)
-
-        result = self.mpicomm.gather(var_values, root=0)
-
-        if (self.cylinder_rank == 0):
-            result = {key: value
-                for dic in result
-                for (key, value) in dic.items()
-            }
-            return result
-
-    def report_var_values_at_rank0(self, header="", print_zero_prob_values=False):
-        """ Pretty-print the values and associated statistics for 
-        non-anticipative variables across all scenarios. """
-
-        var_values = self.gather_var_values_to_rank0(get_zero_prob_values=print_zero_prob_values)
-
-        if self.cylinder_rank == 0:
-
-            if len(header) != 0:
-                print(header)
-
-            scenario_names = sorted(set(x for (x,y) in var_values))
-            max_scenario_name_len = max(len(s) for s in scenario_names)
-            variable_names = sorted(set(y for (x,y) in var_values))
-            max_variable_name_len = max(len(v) for v in variable_names)
-            # the "10" below is a reasonable minimum for floating-point output
-            value_field_len = max(10, max_scenario_name_len)
-
-            print("{0: <{width}s} | ".format("", width=max_variable_name_len), end='')
-            for this_scenario in scenario_names:
-                print("{0: ^{width}s} ".format(this_scenario, width=value_field_len), end='')
-            print("")
-            
-            for this_var in variable_names:
-                print("{0: <{width}} | ".format(this_var, width=max_variable_name_len), end='')
-                for this_scenario in scenario_names:
-                    this_var_value = var_values[this_scenario, this_var]
-                    if (this_var_value == None) and (not print_zero_prob_values):
-                        print("{0: ^{width}s}".format("-", width=value_field_len), end='')
-                    else:
-                        print("{0: {width}.4f}".format(this_var_value, width=value_field_len), end='')
-                    print(" ", end='')
-                print("")
-
-
